Moved to tools/js-tools
