--- tools/eslint/eslint-plugin-jetpack/index.cjs
@@ -1,5 +1,0 @@
-module.exports = {
-	rules: {
-		'use-recommended-jetpack-components': require( './rules/use-recommended-jetpack-components.cjs' ),
-	},
-};
