+++ tools/js-tools/eslint-plugin-jetpack/index.mjs
@@ -1,0 +1,7 @@
+import useRecommendedJetpackComponents from './rules/use-recommended-jetpack-components.mjs';
+
+export default {
+	rules: {
+		'use-recommended-jetpack-components': useRecommendedJetpackComponents,
+	},
+};
