+++ tools/js-tools/package.json
@@ -2,6 +2,9 @@
 	"name": "jetpack-js-tools",
 	"private": true,
 	"description": "Avoid cluttering the monorepo root with random JS script deps by putting them in a workspace. This is intended for simple scripts and third-party scripts that we will only run directly or from the monorepo root.",
+	"scripts": {
+		"test-eslint-plugin-jetpack": "node eslint-plugin-jetpack/tests/use-recommended-jetpack-components.test.mjs"
+	},
 	"bin": {
 		"eslint": "eslint",
 		"eslint-changed": "eslint-changed",
